+++ .dir-locals.el
@@ -1,4 +1,5 @@
 ((nil . ((project-vc-name . "Modeline")
          (wal-project-build-default-cmd . "make dist")
          (wal-project-clean-default-cmd . "make clean")
-         (wal-project-install-default-cmd . "make package-install"))))
+         (wal-project-install-default-cmd . "make package-install")
+         (wal-project-test-default-cmd . "make test && cat coverage/results.txt"))))