+++ test/whale-line-tab-bar-test.el
@@ -1,0 +1,15 @@
+;;; whale-line-tab-bar-test.el --- Tests for custom functionality. -*- lexical-binding: t; -*-
+
+;;; Commentary:
+;;
+;; Tests for the custom functionality.
+
+;;; Code:
+
+(require 'whale-line-tab-bar)
+
+;;; whale-line-tab-bar-test.el ends here
+
+;; Local Variables:
+;; no-byte-compile: t
+;; End: